+++ NEWS.md
@@ -1,5 +1,9 @@
 # hlapro (development version)
 
+* New `get_public()` to lookup the public epitope of an allele in serological 
+  notation, and `get_broad()` to lookup the corresponding broad-level allele 
+  of a serological split.
+
 * New `etrl_lookup()` to get serological equivalents of an allele from the 
   [ETRL HLA tables](https://etrl.eurotransplant.org). These tables are also
   shipped with the package as the `etrl_hla` data frame.